requirements.txt
altgraph==0.17.3
bottle==0.12.23
bottle-websocket==0.2.9
Eel==0.15.1
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
h5py==3.8.0
igorwriter==0.3.0
macholib==1.16.2
mypy==0.991
mypy-extensions==1.0.0
numpy==1.24.1
opencv-python==4.7.0.68
Pillow==9.4.0
polars==0.16.4
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
tinydb==4.7.1
types-Pillow==9.4.0.6
typing_extensions==4.4.0
whichcraft==0.6.1
zope.event==4.6
zope.interface==5.5.2